+++ CHANGELOG.md
@@ -6,9 +6,7 @@ All future updates for CliMenu will be documented in this file
 
 ## [0.3.0] CliMenu Components v1
 ### Added:
-- `Label` and `LabelDisplayer`.
-- `Button`, `ButtonDisplayer` and `ButtonInput`.
-- `Toggle`, `ToggleDisplayer` and `ToggleInput`.
+- Added new components and their behaviors: Label, InputField, Toggle, Bar, Slider and Button.
 
 ### Changed:
 - Now `ComponentConfig`'s constructor sets a default value to `FocusManager` and `LifeCycle` by default.